functions: added adjust_parameters path for stim def get_result: choice of parameters you want to choose def access_results: print a guide on how the result array is indexed def create_dict: returns a result with the selected monitor, and selected variables def plot_tvb_result: automatising the plotting def calculate_PCI def sim_init: change a bit the initialization of the simulator, takes in the parameters argument instead of all the sub-parameters dict, changing the stimulation argument
whole_pipeline: until caclulate pci

nuu_tools : change line 318 the save result to solve the "unhomogeneous third part" error

zerlaut: change numpy.int to int

+++ tvb_model_reference/src/Zerlaut.py
@@ -213,7 +213,7 @@ class Zerlaut_adaptation_first_order(Model):
         doc="""inhibitory decay [ms]""")
 
     N_tot = NArray(
-        dtype=numpy.int,
+        dtype=int,
         label=":math:`N_{tot}`",
         default=numpy.array([10000]),
         domain=Range(lo=1000, hi=50000, step=1000),
@@ -238,14 +238,14 @@ class Zerlaut_adaptation_first_order(Model):
         doc="""fraction of inhibitory cells""")
 
     K_ext_e = NArray(
-        dtype=numpy.int,
+        dtype=int,
         label=":math:`K_ext_e`",
         default=numpy.array([400]),
         domain=Range(lo=0, hi=10000, step=1),  # inhibitory cell number never overcomes excitatory ones
         doc="""Number of excitatory connexions from external population""")
 
     K_ext_i = NArray(
-        dtype=numpy.int,
+        dtype=int,
         label=":math:`K_ext_i`",
         default=numpy.array([0]),
         domain=Range(lo=0, hi=10000, step=1),  # inhibitory cell number never overcomes excitatory ones
@@ -350,7 +350,7 @@ class Zerlaut_adaptation_first_order(Model):
 
     state_variables = 'E I W_e W_i noise'.split()
     _nvar = 5
-    cvar = numpy.array([0], dtype=numpy.int32)
+    cvar = numpy.array([0], dtype=int)
 
     def dfun(self, state_variables, coupling, local_coupling=0.00):
         r"""
